model_checker/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dacite==1.8.1
exceptiongroup==1.2.1
fastapi==0.110.3
flatbuffers==24.3.25
gast==0.5.4
google-auth==2.29.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.63.0
h11==0.14.0
h5py==3.11.0
httptools==0.6.1
idna==3.7
inject==5.2.1
keras==2.8.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
mashumaro==3.13
nats-py==2.4.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
orjson==3.10.1
pillow==10.3.0
protobuf==3.20.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf-estimator-nightly==2.8.0.dev2021122109
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
